media: venus: Rename venus_caps to hfi_plat_caps
[ Upstream commit 8f3b41dcfb9a0fa2d2ca0af51c3eebd670dc153b ]

Now when we have hfi platform make venus capabilities an
hfi platform capabilities.

Stable-dep-of: 9edaaa8e3e15 ("media: venus: hfi_parser: refactor hfi packet parsing logic")
